+++ src/daft-local-execution/src/sources/scan_task.rs
@@ -4,6 +4,9 @@ use daft_io::IOStatsContext;
 use daft_micropartition::MicroPartition;
 use daft_scan::ScanTask;
 use futures::{stream, StreamExt};
+use snafu::ResultExt;
+
+use crate::JoinSnafu;
 
 use super::source::{Source, SourceStream};
 
@@ -20,16 +23,17 @@ impl ScanTaskSource {
 impl Source for ScanTaskSource {
     fn get_data(&self) -> SourceStream {
         log::debug!("ScanTaskSource::get_data");
-        let stream = stream::iter(self.scan_tasks.clone().into_iter().map(|scan_task| {
-            let io_stats = IOStatsContext::new("MicroPartition::from_scan_task");
-            let out =
-                std::thread::spawn(move || MicroPartition::from_scan_task(scan_task, io_stats))
-                    .join()
-                    .expect("Failed to join thread")?;
-
+        let stream = stream::iter(self.scan_tasks.clone().into_iter().map(|scan_task| async {
+            tokio::task::spawn_blocking(move || {
+                let io_stats = IOStatsContext::new("MicroPartition::from_scan_task");
+                MicroPartition::from_scan_task(scan_task, io_stats)
+                    .map(Arc::new)
+                    .map_err(Into::into)
+            })
+            .await
+            .context(JoinSnafu {})?
             // TODO: Implement dynamic splitting / merging of MicroPartition from scan task
-            Ok(Arc::new(out))
         }));
-        stream.boxed()
+        stream.buffered(self.scan_tasks.len()).boxed()
     }
 }
